changed visibilities for proper inheritance of HTML_QuickForm_Renderer_ArraySmarty
There were some methods and property in HTML_QuickForm_Renderer_Array
that could not be overloaded in HTML_QuickForm_Renderer_ArraySmarty
because of the private access level - this lead to an unexpected output
of HTML_QuickForm_Renderer_ArraySmarty.

Also I removed the initialisation of $_elementIdx from
Renderer_Array::startForm() because the prperty was not decelerated nor
used in Renderer_Array and created a Renderer_ArraySmarty::startForm()
which inits $_elementIdx now.